Convert `polyam/packs/admin.jsx` to TS
Glitch-soc did this in the merge.
