+++ app/javascript/flavours/polyam/theme.yml
@@ -1,7 +1,7 @@
 #  (REQUIRED) The location of the pack files.
 pack:
   admin:
-    - packs/admin.jsx
+    - packs/admin.tsx
     - packs/public.jsx
   auth: packs/public.jsx
   common:
